requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
langchain>=0.1.14
langchain-community>=0.0.20
langchain-core>=0.1.30
langchain-huggingface>=0.0.3
transformers>=4.40.0
torch>=2.2.0
sentence-transformers==2.7.0
huggingface-hub>=0.20.0
faiss-cpu>=1.8.0
langfuse>=2.30.0
pymilvus==2.4.0





